Egen/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.1
python-dotenv>=1.0.0 